refactor(core): extract CacheMixin to encapsulate caching state\n\n- Extract `_get_local_cache`, `_update_local_cache`, and `_check_cache_hit` into `CacheMixin`\n- Use lazy initialization for `_cache` inside the mixin methods to avoid leaky abstractions.\n- Use strict type annotations and fix MRO shadowing.
